Update execute.py
I am proposing these changes for a small enhancement of the tool.

While using this tool in PowerFactory, I noticed that by default it uses parallel computing with all the CPU cores in the task automation. I usually feel this is the fastest way to execute long scripts/calculations in PowerFactory, however in some cases the user might need to use some of the CPU for other tasks. This is why I am proposing this change.

I added an extra Input Parameter on the execute.ComPython object in PowerFactory with these characteristics:
Type: int
Name: parallelComp
Value: 1
Unit: 
Description: 1 = Enables parallel computing; 0 = Parallel computing is disabled

Then in the # Read the script options section I have added a line to get this input parameter into the python script:
options.parallelComp : bool = bool(thisScript.GetInputParameterInt('parallelComp')[1])

Finally, before executing the task automation, I have set the value of the parallel computing to the script input variable: 
taskAuto.iEnableParal = options.parallelComp

These two simple lines give an extra feature for the cases where the user might want to have some of the CPU free. In addition in some cases might be of interest to see in the output window all the information/events/warnings/other while running the tool, this is not possible with parallel computation.