Update setupCase.py
createFault:
I have added the options list as an argument. This way, the fault start time is also passed to this function and it can be sent, instead of the default 3seconds.

setupCase:
When setting the stop time of the study case, I have added an if statement to detect if it is a fault case or not. If not it will set the time as before. If it is a fault case, it will set the stop time of the study case according to the fault starting time.