Update execute.py
I have added two lines to check the extra input parameters of the ComPython object in PowerFactory. One for enabling/disabling parallel computation and another one for the fault starting time called:
parallelComp
faultStartTime

I have change the line were the iEnablePral is set to 1 by defualt to the input of the script.